pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["./snds"]

[project]
name = "snds"
version = "0.1.0"
dependencies = [
    "aiohttp==3.9.3",
    "ndn-python-repo==0.3",
    "PyYAML==6.0.1",
    "configparser==7.0.0",
]
